implemented bolding via keywords
Realized I was going down the wrong path by trying to read each individual character and the .FontStyle property. While, that may be needed for reading the current format in the Spreadsheet ... it's tough to keep track of the indices.

For the sake of time and the design requirements, the keyword formatting will suffice since customized formatting would be a "nice-to-have"

CHANGES:
* added `Dim boldArray(20) As String`
** the `20` is an arbitrary magic number for now
* added `IsInArray()` to find strings within a string array.
** thanks to https://stackoverflow.com/q/11109832/5411712
* added code to split the string value of each Cell using the space character as a delimeter .... outputs to cellWordsArr
* added a loop over the cellWordsArr to add the '*' appropriately
** this is the BOLD code to be refactored

TODO:
* refactor BOLD code into a General Format Function that accepts Italics / Underlined / Bold as parameters

* make 2 more arrays called `italicsArray` and `underlinedArray`

* add code to manage merged cells and include {quote} tag appropriately