copypasta no longer fails when "||" missing from first line
also various comment updates.

example of what caused failure:

||Step Name||Description||Expected||Notes
||Step 1..........|| blah.............|| blah........||Notes||

vs what would be expected:

||Step Name||Description||Expected||Notes||
||Step 1..........|| blah.............|| blah........||Notes||

description of the issue:
When "||" was missing from the end of the first line of the JIRA table, the copypasta() function would keep reading until the end of the "Step 1" row, treating the whole thing as one row.

Co-Authored-By: Chris Toh <cetoh@users.noreply.github.com>