added Eric Cole's enhancements
link to JIRA suggestion:
https://jira/jira/browse/TTB-686?filter=40249

Title:
Small enhancements to JTT

Description:
I have modified my script files to offer a few additional features and fix a couple of bugs.

I fixed the issue that generates an error when a quotation mark is in a cell.
I fixed the issue that generates an error when a number only is used in a cell.
I added a feature to allow for auto numbering of steps, preconditions, and VPs.
I added a feature to allow for a double back-slash to be interpreted as a new line break since that is how JIRA reads it.
I am attaching the two files I modified to achieve these results for your code review and to see if you think these could be beneficial.

Auto numbering works if under the Step column, only "precondition", "pc", "step", or "vp" is entered. For instance if you only entered "step" for all steps but had 5 lines of steps and then clicked generate then it would turn those into "Step 1", "Step 2", "Step 3", "Step 4", and "Step 5". I does the same thing for VPs and Preconditions. It will keep track of each of the numbers separately. However for this to work right, all steps, vps, and preconditions have to entered this same way. So if you want to use this feature then all steps should be listed without the numbers.