apply .trim() to all cells
before when a User would paste a Jira Table that was generated by JTT, there would be increasingly more trailing and leading whitespaces. 

now .trim() makes sure that no trailing nor leading whitespaces get into the Spreadsheet output... however due to having to copy directly from the spreadsheet, there are always a small number of tabs in the end result. This is a minor issue that can be resolved by copying from a prompt instead (TODO)